test: split into Functional and Arch
