+++ pest.xml.dist
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:noNamespaceSchemaLocation="vendor/phpunit/phpunit/phpunit.xsd"
+         xsi:noNamespaceSchemaLocation="./vendor/phpunit/phpunit/phpunit.xsd"
          bootstrap="vendor/autoload.php"
          colors="true"
          cacheResultFile=".build/phpunit/result.cache"
@@ -14,20 +14,10 @@
         <bootstrap class="Ergebnis\PHPUnit\SlowTestDetector\Extension"/>
     </extensions>
     <testsuites>
-        <testsuite name="Functional Tests">
-            <directory>tests/src/Functional</directory>
+        <testsuite name="Arch Tests">
+            <directory>tests/src/Arch</directory>
         </testsuite>
     </testsuites>
-    <coverage>
-        <report>
-            <html outputDirectory=".build/coverage"/>
-            <text outputFile=".build/coverage.txt"/>
-            <clover outputFile=".build/logs/clover.xml"/>
-        </report>
-    </coverage>
-    <logging>
-        <junit outputFile=".build/report.junit.xml"/>
-    </logging>
     <source>
         <include>
             <directory>src</directory>
